Move all frontend tests to backend
